refactor: rename user response and user create dto
